Added smtpOutlook function
IF on a machine that has outlook configured, the function creates the Outlook Com Object and then sends emails through Outlook. If no AV installed or updated it will prompt user to allow the sending of emails. On the server you can just runt he smtp server as is.  However an MX record needs to be created for the domain that you are sending to. The domain is what should be put in as the IP variable, however, if you set an stmp_TO it really just tries to send there and the IP doesn't matter then except to get the powershell client to run.