Update ReadMe
Attempted to update the Readme to provide some info for people to know how each file is trying to exfil the data. This information may not be desired to have in the Readme but in a wiki or notes in each client script. I don't know but thought would share as I think it is helpful for people that may use the tool and ensure they know what is really happening or not happening.  Did not comment about any of the specific Malware or Actors modules.  As well as add some info on the SMTPOutlook option and note on the dns_complete.