monior grammar and remove fake data comment
When this was first submitted to add the details or each option, there were some grammatical issues as well as a copy of the statement about Fake Data being made by fakenamegenerator.com.  The fake data comment was a mistake being added, as that is not how the data in egress-assess is created. It is a great option for a source of fake data to work with Egress-Assess but I believe is wrong so should be deleted.